implementacion de tareas por usuario y actualizacion de tests unitarios

+++ src/main/java/com/martin/todo_list/services/TaskService.java
@@ -3,6 +3,7 @@ package com.martin.todo_list.services;
 import com.martin.todo_list.dto.TaskDTO;
 import com.martin.todo_list.entities.Task;
 import com.martin.todo_list.entities.TaskStatus;
+import com.martin.todo_list.entities.User;
 import com.martin.todo_list.repositories.TaskRepository;
 import org.springframework.stereotype.Service;
 
@@ -18,32 +19,32 @@ public class TaskService {
         this.taskRepository = taskRepository;
     }
 
-    public List<TaskDTO> getAllTasks() {
-        List<Task> tasks = taskRepository.findAll();
+    public List<TaskDTO> getAllTasks(User user) {
+        List<Task> tasks = taskRepository.findAllByUser(user);
 
         return tasks.stream()
                 .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
-    public List<TaskDTO> getTasksByStatus(TaskStatus status) {
-        List<Task> tasks = taskRepository.findAllByStatus(status);
+    public List<TaskDTO> getTasksByStatus(User user, TaskStatus status) {
+        List<Task> tasks = taskRepository.findAllByUserAndStatus(user, status);
         return tasks.stream()
                 .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
-    public TaskDTO saveTask(TaskDTO taskDTO){
+    public TaskDTO saveTask(TaskDTO taskDTO, User user) {
         Task task = mapToEntity(taskDTO);
-
+        task.setUser(user);
         Task savedTask = taskRepository.save(task);
 
         return mapToDTO(savedTask);
     }
 
-    public TaskDTO updateTask(Integer id, TaskDTO newTask){
+    public TaskDTO updateTask(Integer id, TaskDTO newTask, User user){
         Task oldTask = taskRepository.findById(id).orElse(null);
-        if (oldTask != null){
+        if (oldTask != null && oldTask.getUser().getId().equals(user.getId())){
             oldTask.setTitle(newTask.getTitle());
             oldTask.setDescription(newTask.getDescription());
             oldTask.setDueDate(newTask.getDueDate());
@@ -55,9 +56,9 @@ public class TaskService {
         return null;
     }
 
-    public TaskDTO updateTaskStatus(Integer id, TaskStatus newStatus){
+    public TaskDTO updateTaskStatus(Integer id, TaskStatus newStatus, User user){
         Task task = taskRepository.findById(id).orElse(null);
-        if (task != null) {
+        if (task != null && task.getUser().getId().equals(user.getId())) {
             task.setStatus(newStatus);
             Task savedTask = taskRepository.save(task);
             return mapToDTO(savedTask);
@@ -65,19 +66,21 @@ public class TaskService {
         return null;
     }
 
-    public boolean deleteTask(Integer id) {
+    public boolean deleteTask(Integer id, User user) {
         Task task = taskRepository.findById(id).orElse(null);
-        if (task != null) {
+        if (task != null && task.getUser().getId().equals(user.getId())) {
             taskRepository.delete(task);
             return true;
         }
         return false;
     }
 
-    public TaskDTO getTaskById(Integer id) {
+    public TaskDTO getTaskById(Integer id, User user) {
         Task task = taskRepository.findById(id).orElse(null);
-        if (task == null) return null;
-        return mapToDTO(task);
+        if (task != null && task.getUser().getId().equals(user.getId())) {
+            return mapToDTO(task);
+        }
+        return null;
     }
 
     private Task mapToEntity(TaskDTO taskDTO) {
